[OHIF Viewer] VLM Analyzer

+++ dbx/pixels/resources/sql/lakebase/CREATE_LAKEBASE_DICOM_FRAMES.sql
@@ -29,8 +29,8 @@
 CREATE TABLE IF NOT EXISTS DICOM_FRAMES (
     filename TEXT NOT NULL,
     frame INTEGER NOT NULL,
-    start_pos INTEGER NOT NULL,
-    end_pos INTEGER NOT NULL,
+    start_pos BIGINT NOT NULL,
+    end_pos BIGINT NOT NULL,
     pixel_data_pos INTEGER NOT NULL,
     PRIMARY KEY (filename, frame)
 );
